teambudgets-front/requirements.txt
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
streamlit==1.45.0
streamlit_authenticator==0.4.2
